Minutes schema refactoring and changed segment content type
- Refactored all minutes schema files into one
- Changed segment content from string array to string in zod schema, mongoose schema and dummyMinutes file
