--- server/src/schemas/segmentSchema.js
@@ -1,8 +1,0 @@
-import { z } from "zod";
-
-const segmentSchema = z.object({
-  name: z.string(),
-  content: z.array(z.string()),
-});
-
-export default segmentSchema;
